+++ src/app/core/@ngrx/app.state.ts
@@ -1,0 +1,9 @@
+import { ProductsState } from './products/products.state';
+import { RouterState } from './router/router.state';
+
+export const productsFeatureKey = 'products';
+export const routerFeatureKey = 'router';
+export interface AppState {
+  [productsFeatureKey]: ProductsState;
+  [routerFeatureKey]: RouterState;
+}